splitterd admin router, tests failing at this point
