+++ build.gradle
@@ -39,6 +39,7 @@ dependencies {
 
     // TESTS
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
+    testImplementation 'org.springframework.security:spring-security-test'
     testImplementation 'io.projectreactor:reactor-test'
     testImplementation 'org.mockito:mockito-core:3.8.0'
 }
