+++ src/main/java/it/sweven/blockcovid/assemblers/UserAssembler.java
@@ -4,8 +4,8 @@ import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.*;
 
 import it.sweven.blockcovid.entities.user.Authority;
 import it.sweven.blockcovid.entities.user.User;
-import it.sweven.blockcovid.routers.AdminRouter;
-import it.sweven.blockcovid.routers.UserRouter;
+import it.sweven.blockcovid.routers.admin.*;
+import it.sweven.blockcovid.routers.user.UserRouter;
 import java.util.Collections;
 import java.util.Objects;
 import java.util.Set;
@@ -36,12 +36,15 @@ public class UserAssembler implements RepresentationModelAssembler<User, EntityM
                 .withRel("change_password"));
     if (authorities.contains(Authority.ADMIN)) {
       userModel.add(
-          linkTo(methodOn(AdminRouter.class).modifyUser(null, entity.getUsername(), null))
+          linkTo(methodOn(AdminModifyUserRouter.class).modifyUser(null, entity.getUsername(), null))
               .withRel("modify_user"));
-      userModel.add(linkTo(methodOn(AdminRouter.class).listUsers(null)).withRel("list_users"));
-      userModel.add(linkTo(methodOn(AdminRouter.class).delete("", "")).withRel("delete_user"));
       userModel.add(
-          linkTo(methodOn(AdminRouter.class).register(null, "")).withRel("register_user"));
+          linkTo(methodOn(AdminListUsersRouter.class).listUsers(null)).withRel("list_users"));
+      userModel.add(
+          linkTo(methodOn(AdminDeleteUserRouter.class).delete("", null)).withRel("delete_user"));
+      userModel.add(
+          linkTo(methodOn(AdminRegistrationRouter.class).register(null, null))
+              .withRel("register_user"));
     }
     clearAuthorities();
     return userModel;
