+++ src/main/java/it/sweven/blockcovid/dto/CredentialChangeRequestForm.java
@@ -1,6 +1,7 @@
 package it.sweven.blockcovid.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import javax.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
@@ -8,8 +9,10 @@ import lombok.Getter;
 @Getter
 public class CredentialChangeRequestForm {
   @JsonProperty("old_password")
+  @NotNull
   private final String oldPassword;
 
   @JsonProperty("new_password")
+  @NotNull
   private final String newPassword;
 }
