+++ src/main/java/it/sweven/blockcovid/routers/admin/AdminModifyUserRouter.java
@@ -1,0 +1,80 @@
+package it.sweven.blockcovid.routers.admin;
+
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
+import it.sweven.blockcovid.assemblers.UserAssembler;
+import it.sweven.blockcovid.dto.CredentialsWithAuthorities;
+import it.sweven.blockcovid.entities.user.User;
+import it.sweven.blockcovid.services.UserService;
+import java.util.Optional;
+import javax.validation.Valid;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.hateoas.EntityModel;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
+
+@RestController
+public class AdminModifyUserRouter implements AdminRouter {
+  private final UserAssembler userAssembler;
+  private final UserService userService;
+
+  @Autowired
+  public AdminModifyUserRouter(UserAssembler userAssembler, UserService userService) {
+    this.userAssembler = userAssembler;
+    this.userService = userService;
+  }
+
+  @PutMapping(
+      value = "/user/{username}/modify",
+      consumes = "application/json",
+      produces = "application/json")
+  @ResponseBody
+  @ApiResponses({
+    @ApiResponse(
+        responseCode = "200",
+        description = "User details successfully updated",
+        content =
+            @Content(
+                mediaType = "application/json",
+                schema = @Schema(implementation = User.class))),
+    @ApiResponse(
+        responseCode = "400",
+        description = "No credentials provided",
+        content = @Content(schema = @Schema(implementation = void.class))),
+    @ApiResponse(
+        responseCode = "401",
+        description = "Invalid authentication token",
+        content = @Content(schema = @Schema(implementation = void.class))),
+    @ApiResponse(
+        responseCode = "403",
+        description = "Method not allowed",
+        content = @Content(schema = @Schema(implementation = void.class))),
+    @ApiResponse(
+        responseCode = "404",
+        description = "Username not found",
+        content = @Content(schema = @Schema(implementation = void.class)))
+  })
+  @PreAuthorize("#submitter.isAdmin()")
+  public EntityModel<User> modifyUser(
+      @AuthenticationPrincipal User submitter,
+      @PathVariable String username,
+      @Valid @RequestBody CredentialsWithAuthorities newCredentials) {
+    User user;
+    try {
+      user = userService.getByUsername(username);
+    } catch (UsernameNotFoundException e) {
+      throw new ResponseStatusException(HttpStatus.NOT_FOUND, "User " + username + " not found");
+    }
+    Optional.ofNullable(newCredentials.getPassword())
+        .ifPresent(pwd -> userService.setPassword(user, pwd));
+    Optional.ofNullable(newCredentials.getAuthorities())
+        .ifPresent(auth -> userService.updateAuthorities(user, auth));
+    return userAssembler.setAuthorities(submitter.getAuthorities()).toModel(user);
+  }
+}
