+++ src/main/java/it/sweven/blockcovid/routers/admin/AdminNewDeskRouter.java
@@ -1,0 +1,86 @@
+package it.sweven.blockcovid.routers.admin;
+
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
+import it.sweven.blockcovid.assemblers.DeskAssembler;
+import it.sweven.blockcovid.dto.DeskInfo;
+import it.sweven.blockcovid.dto.DeskWithRoomName;
+import it.sweven.blockcovid.entities.user.Authority;
+import it.sweven.blockcovid.entities.user.User;
+import it.sweven.blockcovid.exceptions.DeskNotAvailable;
+import it.sweven.blockcovid.exceptions.RoomNotFoundException;
+import it.sweven.blockcovid.services.DeskService;
+import java.util.Set;
+import java.util.stream.Collectors;
+import org.springframework.hateoas.CollectionModel;
+import org.springframework.hateoas.EntityModel;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
+
+@RestController
+public class AdminNewDeskRouter implements AdminRouter {
+  private DeskAssembler deskAssembler;
+  private DeskService deskService;
+
+  public AdminNewDeskRouter(DeskAssembler deskAssembler, DeskService deskService) {
+    this.deskAssembler = deskAssembler;
+    this.deskService = deskService;
+  }
+
+  @PostMapping(
+      value = "rooms/{nameRoom}/desks/new",
+      consumes = "application/json",
+      produces = "application/json")
+  @ResponseBody
+  @ApiResponses({
+    @ApiResponse(responseCode = "200", description = "Desk successfully created"),
+    @ApiResponse(
+        responseCode = "400",
+        description = "Invalid desk parameters",
+        content = @Content(schema = @Schema(implementation = ResponseStatusException.class))),
+    @ApiResponse(
+        responseCode = "401",
+        description = "Invalid authentication token",
+        content = @Content(schema = @Schema(implementation = ResponseStatusException.class))),
+    @ApiResponse(
+        responseCode = "403",
+        description = "Method not allowed",
+        content = @Content(schema = @Schema(implementation = ResponseStatusException.class))),
+    @ApiResponse(
+        responseCode = "404",
+        description = "Requested room doesn't exist",
+        content = @Content(schema = @Schema(implementation = ResponseStatusException.class))),
+    @ApiResponse(
+        responseCode = "409",
+        description = "Desk already exist (with the same id or position)",
+        content = @Content(schema = @Schema(implementation = ResponseStatusException.class)))
+  })
+  public CollectionModel<EntityModel<DeskWithRoomName>> addDesk(
+      @PathVariable String nameRoom,
+      @AuthenticationPrincipal User submitter,
+      @RequestBody Set<DeskInfo> newDesks) {
+    if (!submitter.getAuthorities().contains(Authority.ADMIN))
+      throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+    for (DeskInfo desk : newDesks) {
+      try {
+        deskService.addDesk(desk, nameRoom);
+      } catch (DeskNotAvailable e) {
+        throw new ResponseStatusException(HttpStatus.CONFLICT);
+      } catch (RoomNotFoundException e) {
+        throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+      } catch (IllegalArgumentException e) {
+        throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+      }
+    }
+    return deskAssembler
+        .setAuthorities(submitter.getAuthorities())
+        .toCollectionModel(
+            newDesks.stream()
+                .map(d -> new DeskWithRoomName(d.getId(), nameRoom, d.getX(), d.getY()))
+                .collect(Collectors.toList()));
+  }
+}
