+++ src/main/java/it/sweven/blockcovid/routers/admin/AdminRouter.java
@@ -1,0 +1,6 @@
+package it.sweven.blockcovid.routers.admin;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@RequestMapping("api/admin")
+public interface AdminRouter {}
